kelly-app-v2/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic[email]>=2.10.0
pydantic-settings>=2.6.1
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
python-multipart>=0.0.12
python-dateutil>=2.9.0
pytz>=2024.1
email-validator>=2.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0

